Ignore prompt and add logs on codeMod
