+++ scripts/deno.json
@@ -5,8 +5,7 @@
     "./release": "./release.ts",
     "./update": "./update.run.ts",
     "./upgrade": "./upgrade.ts",
-    "./tailwind-checker": "./tailwindChecker.ts",
-    "./codemod/jsr": "./codemod/jsr.ts"
+    "./tailwind-checker": "./tailwindChecker.ts"
   },
   "publish": {
     "exclude": [
@@ -19,7 +18,7 @@
   },
   "imports": {
     "@cliffy/prompt": "jsr:@cliffy/prompt@^1.0.0-rc.5",
-    "@deco/codemod-toolkit": "jsr:@deco/codemod-toolkit@^0.2.0",
+    "@deco/codemod-toolkit": "jsr:@deco/codemod-toolkit@^0.2.1",
     "@std/flags": "jsr:@std/flags@^0.224.0",
     "@std/fmt": "jsr:@std/fmt@^1.0.0",
     "@std/fs": "jsr:@std/fs@^1.0.1",
